consolidate cli
