+++ src/asli/asli.py
@@ -1,10 +1,8 @@
 """Perform calculations of the Amundsen Sea Low Index"""
 
-import argparse
 import datetime
 import logging
 import os
-import matplotlib.pyplot as plt
 from pathlib import Path
 from typing import Mapping, Union
 import warnings
@@ -478,137 +476,3 @@ class ASLICalculator:
         ]
 
         return plot_lows(da, df, regionbox=ASL_REGION, **kwargs)
-
-
-def _cli_common_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
-    """Parse command-line args common to calculation and plotting."""
-    parser.add_argument(
-        "-d",
-        "--datadir",
-        nargs="?",
-        type=str,
-        default="./data",
-        help="Path to directory in which to put downloaded data. (Default: ./data)",
-    )
-    parser.add_argument(
-        "-m",
-        "--mask",
-        nargs="?",
-        type=str,
-        default="era5_lsm.nc",
-        help="Land-sea mask file path relative to <datadir>. (Default: era5_lsm.nc)",
-    )
-    parser.add_argument(
-        "-o",
-        "--output",
-        type=str,
-        help="Output file path for file, relative to <datadir>.",
-    )
-    parser.add_argument(
-        "msl_files",
-        nargs="*",
-        type=str,
-        help="Path or glob pattern relative to <datadir> for file(s) containing mean sea level pressure.",
-    )  # msl files/pattern
-
-    return parser
-
-
-def _get_cli_plot_args():
-    """Parse command-line arguments for cli_plot()"""
-
-    parser = argparse.ArgumentParser(
-        prog="asli_plot",
-        description="Plot Amundsen sea low with mean sea level pressure fields.",
-    )
-    parser.add_argument(
-        "-i",
-        "--input",
-        nargs="?",
-        type=str,
-        help="Input CSV file, relative to <datadir>.",
-    )
-    parser.add_argument(
-        "-y",
-        "--year",
-        nargs="?",
-        type=int,
-        help="When present, plot only the year specified",
-    )
-    parser = _cli_common_args(parser)
-
-    return parser.parse_args()
-
-
-def _get_cli_calc_args():
-    """Parse command-line arguments for cli_calc()"""
-
-    parser = argparse.ArgumentParser(
-        prog="asli_calc",
-        description="Calculates the Amundsen Sea Low from mean sea level pressure fields.",
-    )
-    parser = _cli_common_args(parser)
-    parser.add_argument(
-        "-e",
-        "--era5t",
-        action="store_true",
-        help="When present, this flag enables the inclusion of ERA5T initial release data as well as finalised ERA5 data.",
-    )
-    parser.add_argument(
-        "-n",
-        "--numjobs",
-        nargs="?",
-        type=int,
-        default=1,
-        help="Number of processes used by joblib in parallel calculation.",
-    )
-    parser.add_argument(
-        "-M",
-        "--minima",
-        type=int,
-        nargs="?",
-        default=1,
-        help="Max number of minima to locate in pressure field per time step.",
-    )
-
-    return parser.parse_args()
-
-
-def _cli_plot():
-    """Command-line interface to ASLI plotting."""
-
-    args = _get_cli_plot_args()
-
-    a = ASLICalculator(args.datadir, args.mask, args.msl_files[0])
-    a.read_mask_data()
-    a.read_msl_data()
-    # Perform the calculation if no input file is provided
-    if args.input:
-        a.import_from_csv(args.input)
-    else:
-        a.calculate()
-    # Plot all if no specific year is provided
-    if args.year:
-        a.plot_region_year(args.year)
-    else:
-        a.plot_region_all()
-    if args.output:
-        plt.savefig(os.path.join(args.datadir, args.output))
-
-
-def _cli_calc():
-    """Command-line interface to ASL calculation."""
-
-    args = _get_cli_calc_args()
-
-    a = ASLICalculator(args.datadir, args.mask, args.msl_files[0])
-    a.read_mask_data()
-    a.read_msl_data(include_era5t=args.era5t)
-    a.calculate(args.numjobs, num_minima=args.minima)
-
-    if args.output:
-        a.to_csv(args.output)
-
-
-if __name__ == "__main__":
-    _cli_calc()
